// next-leaflet/src/types/payload-types.ts
export interface Job {
  id: number;
  jobName: string;
  providerName?: string | null;
  pool?: string | null;
  URL?: string | null;
  proxy?: string | null;
  type?: string | null;
  status?: ('running' | 'pending' | 'stopped' | 'failed') | null;
  threads?: number | null;
  lastHeartbeat?: string | null;
  updatedAt: string;
  createdAt: string;
}